move delete setting to better location, refresh music tab when changing setting
